+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/Json.kt
@@ -26,8 +26,12 @@ internal object JsonCredentialSerializer {
 
 }
 
-val jsonSerializer by lazy {
-    Json(from = at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer) {
+val vckJsonSerializer by lazy {
+    Json {
+        prettyPrint = false
+        encodeDefaults = false
+        classDiscriminator = "type"
+        ignoreUnknownKeys = true
         serializersModule = SerializersModule {
             polymorphic(CredentialSubject::class) {
                 subclass(AtomicAttribute2023::class)